requirements.txt
python-telegram-bot==20.1
python-dotenv==1.0.0
aiohttp==3.11.11
python-i18n==0.3.9